Move Dropout forward implementation in inference project.
